pyproject.toml
[tool.poetry]
name = "data_ingestion"
version = "0.1.0"
description = "Data Ingestion"
authors = ["Rahsheeda McRae <rahsheeda.a.mcrae@gmail.com>"]
readme = "README.md"
packages = [{include = "data_ingestion"}]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.31.0"
boto3 = "^1.34.91"
python-dotenv = "^1.0.1"
pandas = "^1.1"
awswrangler = "^3.7.3"
urllib3 = ">=1.25.4,<1.27"
ratelimit = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"